Revert "Version 0.6"
This reverts commit 2b354068db6a05ea1474735a3ae7558840f4c3ae.

+++ Levels/Levels.Exits.cs
@@ -281,7 +281,7 @@ namespace FF3LE
         {
             if (updatingLevel) return;
 
-            exits.DestinationShort = (ushort)exitShortDestination.SelectedIndex;
+            exits.DestinationShort = (byte)exitShortDestination.SelectedIndex;
         }
         private void exitDestinationXCoord_ValueChanged(object sender, EventArgs e)
         {
@@ -443,7 +443,7 @@ namespace FF3LE
         {
             if (updatingLevel) return;
 
-            exits.DestinationLong = (ushort)exitLongDestination.SelectedIndex;
+            exits.DestinationLong = (byte)exitLongDestination.SelectedIndex;
         }
         private void exitLongDestinationXCoord_ValueChanged(object sender, EventArgs e)
         {
